Added GetAverageBookReadingPeriod

+++ src/main/java/DAO/ElementDAO.java
@@ -6,7 +6,7 @@ import java.util.Optional;
 public interface ElementDAO<E> {
     void add(E e);
     void update(E e);
-    E get(long eID);
+    E get(int eID);
     //E get(long eID);
     Collection<E> getAll();
     void delete(E e);
